% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_raw.R
\name{backup_raw}
\alias{backup_raw}
\title{Backup Raw Data}
\usage{
backup_raw(from_dir, to_dir, cache_dir = "cache", force = FALSE)
}
\arguments{
\item{from_dir}{Directory to copy from.}

\item{to_dir}{Directory to copy to.}

\item{cache_dir}{Directory to store cached file snapshots.}

\item{force}{Logical indicating if copying should be forced.}
}
\value{
Invisible - this function is used for its side effects
}
\description{
This function copies raw data from an external location (i.e. a mapped network drive)
to a specified local project directory. In addition, a snapshot of the files
being copied is taken and stored in the specified "cache_dir" so that files
are only re-copied if changed.
}
\examples{

\dontrun{
library(jimstools)
library(fs)

network_path <- fs::path(
"H:\\\\ATLRFI", "INDUSTRY", "Rates - Loss Costs", "Exposure_Adjustment_Template",
"2019", "Backup", "Excess Loss Factors", "Raw Data"
)

local_path <- fs::path("data-raw", "xs-loss-factors")

backup_raw(network_path, local_path)
}

}
