% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-docs-badge.R
\name{badge}
\alias{badge}
\title{Create README badges via img.shields.io}
\usage{
badge(label, value, color, url)
}
\arguments{
\item{label}{badge label}

\item{value}{badge value}

\item{color}{badge color}

\item{url}{url}
}
\value{
character string to be pasted into README
}
\description{
Create README badges via img.shields.io
}
