% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{caching}
\alias{caching}
\alias{write_cache}
\alias{read_cache}
\title{Caching Utility Functions}
\usage{
write_cache(x, name = NULL, cache_dir = "cache", overwrite = TRUE)

read_cache(x, name = NULL, cache_dir = "cache")
}
\arguments{
\item{x}{object to cache}

\item{name}{name of object to read in.}

\item{cache_dir}{path to cache directory.}

\item{overwrite}{logical (default = TRUE)}
}
\value{
invisibly returns object passed to the function.

x

invisibly attaches object to parent global environment
}
\description{
Quick caching utility read/write functions utilizing the \code{qs} package.
}
\examples{
mydata <- mtcars
write_cache(mydata) # will save to 'cache/mydata'.
write_cache(mydata, "mydata-v2", cache_dir = "data/temp") # will save to 'data/temp/mydata-v2'

# read back in
read_cache(mydata)
}
\seealso{
\code{\link[qs:qread]{qs::qread()}} and \code{\link[qs:qsave]{qs::qsave()}}
}
