% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-coalesce_join.R
\name{coalesce_join}
\alias{coalesce_join}
\title{Coalesce Join}
\usage{
coalesce_join(
  x,
  y,
  by = NULL,
  suffix = c(".x", ".y"),
  join = dplyr::full_join,
  ...
)
}
\arguments{
\item{x}{data.frame}

\item{y}{data.frame to join with}

\item{by}{variables to join by}

\item{suffix}{common variables suffix}

\item{join}{type of join}

\item{...}{passed to join function}
}
\value{
joined data.frame with values replaced
}
\description{
Coalesce Join
}
