% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-text.R
\name{collapseRows}
\alias{collapseRows}
\title{Collapse Rows}
\usage{
collapseRows(df, variable)
}
\arguments{
\item{df}{a data frame}

\item{variable}{group variable to be collapsed}
}
\value{
a data frame with an updated \code{variable} column
}
\description{
for each group, sets the top row of the group to the group's value.  All other
rows in the group are set to "".  See the example
}
