% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actuary-compare_to_prior.R
\name{compare_to_prior}
\alias{compare_to_prior}
\title{Compare to Prior}
\usage{
compare_to_prior(data, col)
}
\arguments{
\item{data}{current data}

\item{col}{columns to compare (character vector)}
}
\value{
new data frame with added columns for changes and priors
}
\description{
Compares data across evaluations.
}
\keyword{actuarial}
