% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{Create Project}
\usage{
create_project(
  path,
  rstudio = rstudioapi::isAvailable(),
  open = rlang::is_interactive(),
  fields = list()
)
}
\arguments{
\item{path}{Path}

\item{rstudio}{Boolean - Use RStudio?}

\item{open}{Boolean - Open?}

\item{fields}{passed to \code{\link[usethis:use_description]{usethis::use_description()}}'s \code{fields} argument.}
}
\value{
RStudio project (side-effect)
}
\description{
Create Project
}
