% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_rstudio_project}
\alias{create_rstudio_project}
\title{Create RStudio Project}
\usage{
create_rstudio_project(name, path = ".", ...)
}
\arguments{
\item{name}{name of project}

\item{path}{defaults to "."}

\item{...}{passed to \code{\link[=render_template]{render_template()}}'s \code{data} argument}
}
\value{
same as \code{\link[usethis:create_package]{usethis::create_package()}}.
}
\description{
Create RStudio Project
}
