% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs.R
\name{cs}
\alias{cs}
\title{Vector of character strings from list of unquoted names}
\usage{
cs(...)
}
\description{
Returns a vector of character strings from list of unquoted names.
(Copied directly from the Hmisc package).
}
\examples{
cs(a,cat,dog)
}
