% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{default_colors}
\alias{default_colors}
\title{Vectors of colors for figures}
\usage{
default_colors()
}
\value{
Vector of character strings representing colors, in RGB.
}
\description{
Creates different vectors of related colors that may be useful for figures.
}
\examples{
plot(1:4,rep(1,4),col=default_colors(),pch=19,cex=5)
plot(1:16,rep(1,16),col=default_colors(),pch=19,cex=5)
}
\keyword{utilities}
