% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-download_bttn.R
\name{download_bttn}
\alias{download_bttn}
\title{Download Button}
\usage{
download_bttn(outputId, label = "Download", class = NULL, icon = NULL, ...)
}
\arguments{
\item{outputId}{output id}

\item{label}{label}

\item{class}{class}

\item{icon}{icon}

\item{...}{passed to tags$a}
}
\value{
download button HTML
}
\description{
Download button without opening up new window in browser when downloaded.
}
