% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_from_vars.R
\name{expand_from_vars}
\alias{expand_from_vars}
\title{Create a Skeleton 'Grid' for Data based on combinations of unique column values}
\usage{
expand_from_vars(data, vars)
}
\arguments{
\item{data}{data.frame}

\item{vars}{variables to use}
}
\value{
expanded data.frame
}
\description{
Create a Skeleton 'Grid' for Data based on combinations of unique column values
}
