% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{get_os}
\alias{get_os}
\alias{is_32bit_os}
\alias{is_64bit_os}
\title{Determine Operating System (OS)}
\usage{
get_os()

is_32bit_os()

is_64bit_os()
}
\value{
OS name in lower case: windows, mac, linux, etc.
}
\description{
Determine the operating system (OS) of your machine.
}
\examples{
get_os()
}
\seealso{
https://www.r-bloggers.com/identifying-the-os-from-r/
}
\concept{system}
