% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actuary-group_by_occurrence.R
\name{group_by_occurrence}
\alias{group_by_occurrence}
\title{group_by_occurrence}
\usage{
group_by_occurrence(dat)
}
\arguments{
\item{dat}{loss_data by claim ID}
}
\value{
data grouped by occurrence
}
\description{
Group loss data by occurrence ID.
}
\examples{
\dontrun{
dat <- readRDS("data/shiny-data.rds") \%>\%
  filter(
    eval_dt == as.Date("2018-12-31"),
    member == "Ace Endico Corporation"
  )
test <- by_occurrence(dat)
}


}
\keyword{actuarial}
