% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-download_bttn.R, R/shiny-valid_email.R
\name{is_valid_email}
\alias{is_valid_email}
\title{Ensures a text string is a valid email address}
\usage{
is_valid_email(x)

is_valid_email(x)
}
\arguments{
\item{x}{string}
}
\value{
logical

logical
}
\description{
Ensures a text string is a valid email address

Ensures a text string is a valid email address
}
