% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_filt.R
\name{multi_filt}
\alias{multi_filt}
\title{Multi Filt}
\usage{
multi_filt(data, vars, filts)
}
\arguments{
\item{data}{data}

\item{vars}{variables}

\item{filts}{filters}
}
\value{
filtered df
}
\description{
Filter across multiple variables.
}
