% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-operators.R
\name{operators}
\alias{operators}
\alias{\%notin\%}
\alias{\%allin\%}
\alias{\%anyin\%}
\alias{\%nonein\%}
\alias{\%partin\%}
\title{Operators}
\usage{
x \%notin\% vector

x \%allin\% vector

x \%anyin\% vector

x \%nonein\% vector

pattern \%partin\% vector
}
\arguments{
\item{x}{Numeric or character vector.}

\item{vector}{Numeric or character vector.}

\item{pattern}{Character string containing \strong{regular expressions} to be matched.}
}
\value{
logical

logical
}
\description{
Extensions of base operators.

inverse of \code{\%in\%}
}
\examples{
data <- data.frame(
  id = 1:10,
  x = as.integer(runif(10, 0, 10))
)

data[data$id \%notin\% c(1, 3, 5, 7, 9), ]

1:2 \%allin\% 1:3  # TRUE
3:4 \%allin\% 1:3  # FALSE
3:4 \%anyin\% 1:3  # TRUE
4:5 \%anyin\% 1:3  # FALSE
3:4 \%nonein\% 1:3  # FALSE
4:5 \%nonein\% 1:3  # TRUE
"Bei" \%partin\% c("Beijing", "Shanghai")  # TRUE
"bei" \%partin\% c("Beijing", "Shanghai")  # FALSE
"[aeiou]ng" \%partin\% c("Beijing", "Shanghai")  # TRUE

}
\seealso{
\code{\link[base:any]{base::any()}}, \code{\link[base:match]{base::match()}}, and \code{\link[base:all]{base::all()}}

\code{\link[base:grep]{base::grepl()}}
}
\keyword{operators}
