% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get_rstudio_projects.R
\name{parse_proj_path}
\alias{parse_proj_path}
\title{#' @importFrom readr read_lines
#' @importFrom stringr str_subset
get_rsudio_projects <- function() {
file <- path
projs <- readr::read_lines(file)
stringr::str_subset(projs, "^\\s*$", negate = TRUE)
}}
\usage{
parse_proj_path(proj_path)
}
\description{
#' @importFrom readr read_lines
#' @importFrom stringr str_subset
get_rsudio_projects <- function() {
file <- path
projs <- readr::read_lines(file)
stringr::str_subset(projs, "^\\s*$", negate = TRUE)
}
}
