% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_winslash.R
\name{paste_winslash}
\alias{paste_winslash}
\title{Paste Winslash - Addin Function}
\usage{
paste_winslash()
}
\value{
normalized path character string
}
\description{
Paste Winslash - Addin Function
}
\details{
assign to keyboard shortcut Ctrl + Shift + P for easy access.
}
