% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_date.R
\name{print_date}
\alias{print_date}
\title{Print the current date in a pretty format}
\usage{
print_date()
}
\value{
string
}
\description{
Print the current date in a pretty format
}
\examples{
print_date()
}
