% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_unique.R
\name{pull_unique}
\alias{pull_unique}
\title{Pull all unique values for a variable}
\usage{
pull_unique(data, var, sort = TRUE, decreasing = FALSE, names = TRUE)
}
\arguments{
\item{data}{data.frame}

\item{var}{variable name}

\item{sort}{logical (default = TRUE)}

\item{decreasing}{logical (default = FALSE)}

\item{names}{logical (default = TRUE)}
}
\value{
vector
}
\description{
Pull all unique values for a variable
}
