% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_tbl.R
\name{quick_table}
\alias{quick_table}
\title{htable
Create a quick html table using my favorites options for \code{\link[DT:datatable]{DT::datatable()}}}
\usage{
quick_table(df, ...)
}
\arguments{
\item{df}{data.frame to display}

\item{...}{options sent through to \link[DT:datatable]{DT::datatable}}
}
\description{
htable
Create a quick html table using my favorites options for \code{\link[DT:datatable]{DT::datatable()}}
}
\examples{
quick_table(mtcars)
}
