% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{read_cache}
\alias{read_cache}
\title{Read Cache}
\usage{
read_cache(name, cache_dir = "cache")
}
\arguments{
\item{name}{name of object to read in.}

\item{cache_dir}{path to cache directory.}
}
\value{
invisibly attaches object to parent global environment
}
\description{
Quick caching utility read function.
}
