% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{render_template}
\alias{render_template}
\title{Render Template}
\usage{
render_template(template_path, out_path, data = list(), open = TRUE)
}
\arguments{
\item{template_path}{Path to template}

\item{out_path}{Path to output the rendered template}

\item{data}{list to pass to \code{\link[whisker:whisker.render]{whisker::whisker.render()}}}

\item{open}{boolean - open template after it renders?}
}
\value{
invisible
}
\description{
Renders a template using \code{\link[whisker:whisker.render]{whisker::whisker.render()}} for templates
used in a fashion similar to \code{\link[usethis:use_template]{usethis::use_template()}}.
}
