% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-browser-search.R
\name{search}
\alias{search}
\alias{search_online}
\alias{search_gh}
\alias{search_rseek}
\alias{search_finzi}
\alias{search_nabble}
\alias{search_google}
\alias{search_rproject}
\alias{search_metacran}
\title{Generalized Search and Search Engine Examples}
\usage{
search_online(s, query_url = "https://google.com/search?q=")

search_gh(
  s,
  type = "all",
  language = NULL,
  topic = NULL,
  user = NULL,
  org = NULL
)

search_rseek(s)

search_finzi(s)

search_nabble(s)

search_google(s)

search_rproject(s)

search_metacran(s)
}
\arguments{
\item{s}{string to search for}

\item{query_url}{string representing the URL to query; defaults to Google}

\item{type}{what to search for, see details for options}

\item{language}{optional language filter}

\item{topic}{optional topic filter}

\item{user}{optional user filter}

\item{org}{optional org filter}
}
\description{
Use these functions to search the web directly from your browser,
building advanced queries and supplying common useful R related domains.

Query Github's internal search engine.

Query a search on \href{https://rseek.org/}{rseek.org}.

Query a search on the R Nabble Forum. Nabble is an innovative search engine
for R messages.
}
\details{
See \url{https://cloud.r-project.org/search.html} which showcases Google's
advanced search feature to query only R-Project domain sites via the
\href{http://www.google.com/advanced_search}{Google Search Engine}.
}
\section{Functions}{
\itemize{
\item \code{search_gh}: 

\item \code{search_rseek}: 

\item \code{search_finzi}: 

\item \code{search_nabble}: 

\item \code{search_google}: 

\item \code{search_rproject}: 

\item \code{search_metacran}: 
}}

\examples{
# default search on google
search_online("polished.tech")

# search github (note: &ref=opensearch)
search_online("polished", "https://github.com/search?q=\%s&ref=opensearch")

# search Github with language:r, org:tychobra for 'polished' (note: '\%3A' represents a ':')
search_online("polished", "https://github.com/search?q=\%s+language\%3Ar+org\%3Atychobra")
search_gh("websocket", language = "r", topic = "rshiny")

# search your org
search_gh("polished", org = "tychobra")
}
\references{
\itemize{
\item \url{http://www.sashagoodman.com/}
}

\itemize{
\item \url{http://finzi.psych.upenn.edu/search.html}
\item \url{http://finzi.psych.upenn.edu/search/manual.html#query}
}

\itemize{
\item \url{https://cloud.r-project.org/search.html}
\item \url{http://n4.nabble.com/help/Answer.jtp?id=31}
}
}
\keyword{search_engines}
