% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-set_wd.R
\name{set_wd}
\alias{set_wd}
\title{Set working directory to the path of \strong{current} script}
\usage{
set_wd(dir = NULL)
}
\arguments{
\item{dir}{\code{NULL} (default) or a character string specifying the working directory.

If \code{NULL}, set working directory to the path of \strong{the current R script}.}
}
\description{
Set working directory to the path of \strong{current} script
}
\examples{
\dontrun{
set_wd()  # set working directory to the path of the current R script

set_wd("D:/")  # "\" is not allowed, you should use "/"

set_wd("../")  # set working directory to the parent directory
}
}
\seealso{
\code{\link{setwd}}
}
