% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-startup.R
\name{startup}
\alias{startup}
\alias{editrprof}
\alias{editrenv}
\title{R Startup Functions}
\usage{
editrprof(scope = c("user", "project"))

editrenv(scope = c("user", "project"))
}
\arguments{
\item{scope}{}
}
\value{
invisible
}
\description{
Use these utility functions to quickly edit your \code{.Rprofile},
\code{.Renviron}, or other startup files.
}
\seealso{
See \code{\link[base:Startup]{base::Startup()}} for details on the R specific startup process.
}
