% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_linguist.R
\name{use_gh_linguist}
\alias{use_gh_linguist}
\title{GitHub Linguist Exclusions}
\usage{
use_gh_linguist(excludes = c("*.js", "*.html", "*.css"))
}
\arguments{
\item{excludes}{Character vector of wildcards indicating extensions to be labelled as \code{linguist-vendored} and ignored by GitHub's language detector \code{github-linguist}.}
}
\value{
invisibly returns the text written to \code{.gitattributes}
}
\description{
This function creates a \code{.gitattributes} file in the root of your project in
order to exclude certain file types from being detected as a certain programming language on GitHub. See details for more information.
}
\details{
By default this function will write \verb{*.js}, \verb{*.html}, and \verb{*.css} to a \code{.gitattributes} file in order for \code{github-linguist} to ignore Javascript, HTML, and CSS languages in your repositories language section.

The resulting file will look like so:\preformatted{# github linguist exclusions:
*.js linguist-vendored
*.html linguist-vendored
*.css linguist-vendored
}
}
\examples{
\dontrun{
use_gh_linguist()
}
}
