% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{user_info}
\alias{user_info}
\title{Get User-Related Information}
\usage{
user_info()
}
\value{
invisibly returns \code{os_info} vector
}
\description{
Get user-related information.
}
\examples{
user_info()

}
