% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_list.R
\name{view_list}
\alias{view_list}
\title{view_list}
\usage{
view_list(list, ...)
}
\arguments{
\item{...}{options sent through to \code{\link[listviewer:jsonedit]{listviewer::jsonedit()}}}

\item{l}{list to display}
}
\description{
Interactive view of list structures with \code{\link[listviewer:jsonedit]{listviewer::jsonedit()}}
}
\examples{
view_list(mtcars)
}
