% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{write_cache}
\alias{write_cache}
\title{Write Cache}
\usage{
write_cache(x, name = NULL, cache_dir = "cache", overwrite = TRUE)
}
\arguments{
\item{x}{object to cache}

\item{name}{name to store object with}

\item{cache_dir}{path to cache directory}

\item{overwrite}{logical (default = TRUE)}
}
\value{
invisible
}
\description{
Quick caching utility write function.
}
