% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs.R
\name{cs}
\alias{cs}
\title{Vector of character strings from list of unquoted names}
\usage{
cs(...)
}
\arguments{
\item{...}{dots to quote and \code{c()}.

Returns a vector of character strings from list of unquoted names.
(Copied directly from the Hmisc package).}
}
\description{
Vector of character strings from list of unquoted names
}
\examples{
cs(a,cat,dog)
}
