% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rstudio_projects.R
\name{get_rstudio_projects}
\alias{get_rstudio_projects}
\title{Get RStudio Projects}
\usage{
get_rstudio_projects()
}
\value{
data.frame
}
\description{
Pulls projects list from RStudio's AppData files.
}
