% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinytools.R
\name{is_valid_email}
\alias{is_valid_email}
\title{Ensures a text string is a valid email address}
\usage{
is_valid_email(x)
}
\arguments{
\item{x}{string}
}
\value{
logical
}
\description{
Ensures a text string is a valid email address
}
