% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-package.R
\docType{package}
\name{jimstools-package}
\alias{jimstools}
\alias{jimstools-package}
\title{jimstools: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimbrig2011/jimstools}
  \item Report bugs at \url{https://github.com/jimbrig2011/jimstools/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@oliverwyman.com}

}
\keyword{internal}
